using latest mock deps, move select-modal as new atom
