+++ packages/@coorpacademy-components/src/atom/select-modal/test/fixtures/colored.tsx
