+++ packages/@coorpacademy-components/src/atom/select-modal/test/fixtures/focused.tsx
@@ -1,4 +1,4 @@
-import {Props} from '../../index.native';
+import {Props} from '../../../select-modal/index.native';
 import {createSelectChoice} from '../../../../hoc/modal/select/test/fixtures/default';
 
 type Fixture = {props: Props};
@@ -9,9 +9,9 @@ const items = select.items || [];
 const fixture: Fixture = {
   props: {
     questionType: 'template',
-    placeholder: 'Feel free to edit',
     analyticsID: 'foo',
-    values: items
+    values: items,
+    isFocused: true
   }
 };
