+++ packages/@coorpacademy-components/src/atom/select-modal/test/fixtures/with-value.tsx
