+++ packages/@coorpacademy-components/src/atom/select-modal/test/on-blur.tsx
@@ -1,22 +1,26 @@
 import test from 'ava';
 import React from 'react';
 import {render, fireEvent} from '@testing-library/react-native';
-import Select from '../index.native';
+import Select from '../../select-modal/index.native';
 import {createSelectChoice} from '../../../hoc/modal/select/test/fixtures/default';
 import mockMobileContext from '../../../test/helpers/mock-mobile-context';
 import {TemplateContext} from '../../../template/app-review/template-context';
+import {ANALYTICS_EVENT_TYPE} from '../../../variables/analytics';
 
-test('should handle on change', t => {
+test('should handle on blur', t => {
+  const analyticsID = 'fake-analytics-id';
   const context = mockMobileContext({
-    logEvent: () => t.pass()
+    logEvent: (eventName: string, options: {id: string}) => {
+      t.is(eventName, ANALYTICS_EVENT_TYPE.CLOSE_SELECT);
+      t.deepEqual(options, {id: analyticsID});
+    }
   });
 
   const select = createSelectChoice({name: 'sel456'});
   const items = select.items || [];
 
-  const analyticsID = 'foo';
   const questionType = 'template';
-  const handleChange = value => {
+  const handleBlur = value => {
     t.is(value, 'App Store');
   };
 
@@ -25,7 +29,7 @@ test('should handle on change', t => {
       <Select
         values={items}
         value={items[1].text}
-        onChange={handleChange}
+        onBlur={handleBlur}
         questionType={questionType}
         analyticsID={analyticsID}
         placeholder="Foo bar baz"
