+++ packages/@coorpacademy-components/src/atom/text/index.native.tsx
@@ -1,10 +1,10 @@
 import * as React from 'react';
-import {Text as TextBase, Platform, StyleSheet, TextStyle} from 'react-native';
+import {Text as TextBase, Platform, StyleSheet, ViewStyle} from 'react-native';
 
 export type Props = {
   // copied from node_modules/react-native/Libraries/Text/TextProps.js
   children: React.ReactNode;
-  style?: TextStyle;
+  style?: ViewStyle | ViewStyle[];
   testID?: string;
   numberOfLines?: number;
 };
